+++ relay-server/src/services/processor/dynamic_sampling.rs
@@ -237,7 +237,6 @@ mod tests {
     use bytes::Bytes;
     use relay_base_schema::events::EventType;
     use relay_base_schema::project::{ProjectId, ProjectKey};
-    use relay_dynamic_config::GlobalConfig;
     use relay_dynamic_config::{MetricExtractionConfig, TransactionMetricsConfig};
     use relay_event_schema::protocol::{EventId, LenientString};
     use relay_protocol::RuleCondition;
@@ -437,11 +436,7 @@ mod tests {
                 event: Annotated::from(event),
                 metrics: Default::default(),
                 sample_rates: None,
-                extracted_metrics: ProcessingExtractedMetrics::new(
-                    project_state.clone(),
-                    Arc::new(GlobalConfig::default()),
-                    envelope.dsc(),
-                ),
+                extracted_metrics: ProcessingExtractedMetrics::new(),
                 config: config.clone(),
                 project_state,
                 sampling_project_state: None,
@@ -713,11 +708,7 @@ mod tests {
             spans_extracted: false,
             metrics: Default::default(),
             sample_rates: Default::default(),
-            extracted_metrics: ProcessingExtractedMetrics::new(
-                project_info.clone(),
-                Arc::new(GlobalConfig::default()),
-                envelope.dsc(),
-            ),
+            extracted_metrics: ProcessingExtractedMetrics::new(),
             config: Arc::new(Config::default()),
             project_state: project_info,
             sampling_project_state: {