+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Update definitions for user agent parsing. ([#3951](https://github.com/getsentry/relay/pull/3951))
 - Extend project config API to be revision aware. ([#3947](https://github.com/getsentry/relay/pull/3947))
 - Removes `processing.max_secs_in_past` from the main config in favor of event retention from the project config. ([#3958](https://github.com/getsentry/relay/pull/3958))
+- Remove support for extrapolation of metrics. ([#3969](https://github.com/getsentry/relay/pull/3969))
 
 **Internal**: